feat(write page): add small batch rules
