+++ web/locales/en/pages/contribute/write.ftl
@@ -10,7 +10,8 @@ citation-input-placeholder = Reference the source of your sentence (required)
 citation =
     .label = Citation
 sc-write-submit-confirm = I confirm that this sentence is <wikipediaLink>public domain</wikipediaLink> and I have permission to upload it.
-sc-review-write-title = What sentences can I add?
+sc-review-single-write-title = What sentences can I add?
+sc-review-small-batch-title = How to add multiple sentences
 new-sentence-rule-1 = <noCopyright>No copyright</noCopyright> restrictions (<cc0>cc-0</cc0>)
 new-sentence-rule-2 = Fewer than 15 words per sentence
 new-sentence-rule-3 = Use correct grammar
@@ -19,6 +20,7 @@ new-sentence-rule-5 = No numbers and special characters
 new-sentence-rule-6 = No foreign letters
 new-sentence-rule-7 = Include appropriate citation
 new-sentence-rule-8 = Ideally natural and conversational (it should be easy to read the sentence)
+login-instruction-multiple-sentences = Login or sign up to add multiple sentences
 how-to-cite = How do I cite?
 how-to-cite-explanation-bold = Cite with a URL link or the full name of the work.
 how-to-cite-explanation = If it’s your own words, just say <italicizedText>“Self Citation”</italicizedText>. We need to know where you found this content so that we can check it is in the public domain and no copyright restrictions apply. For more information about citation see our <guidelinesLink>Guidelines page</guidelinesLink>.
@@ -95,3 +97,9 @@ too-many-files = Too many files
 ## SMALL BATCH SUBMISSION
 small-batch-instruction = <icon></icon> Add multiple public domain sentences
 multiple-sentences-error = You cannot add multiple sentences for a single submission
+small-batch-sentences-rule-1 = Follow guidelines from “What sentences can I add?”
+small-batch-sentences-rule-2 = Add one sentence per line
+small-batch-sentences-rule-3 = Separate sentences into one line by pressing “Enter” or “Return” once
+small-batch-sentences-rule-4 = Add up to 1,000 sentences
+small-batch-sentences-rule-5 = All sentence must have the same domain
+small-batch-sentences-rule-6 = All sentences must have the same citation
