+++ web/src/components/pages/contribution/sentence-collector/write/sentence-write/index.tsx
@@ -56,6 +56,7 @@ export const SentenceWrite: React.FC<Props> = ({
           variantTokens={allVariants}
           selectedVariant={sentenceWriteState.sentenceVariant}
           instructionLocalizedId={instructionLocalizedId}
+          mode={mode}
         />
       </div>
